requirements.txt
pymongo
dnspython
python-dotenv
yfinance
pandas
tqdm
